+++ src/test/java/dev/marksman/enhancediterables/ImmutableNonEmptyFiniteIterableTest.java
@@ -467,6 +467,25 @@ class ImmutableNonEmptyFiniteIterableTest {
     }
 
     @Nested
+    @DisplayName("reduceRight")
+    class ReduceRight {
+
+        @Test
+        void throwsOnNullOperator() {
+            NonEmptyFiniteIterable<Integer> ints = immutableNonEmptyFiniteIterable(1, asList(2, 3));
+            //noinspection ConstantConditions
+            assertThrows(NullPointerException.class, () -> ints.reduceRight(null));
+        }
+
+        @Test
+        void onSize5() {
+            NonEmptyFiniteIterable<String> items = immutableNonEmptyFiniteIterable("1", asList("2", "3", "4", "5"));
+            assertEquals("5,4,3,2,1", items.reduceRight((x, acc) -> acc + "," + x));
+        }
+
+    }
+
+    @Nested
     @DisplayName("reverse")
     class Reverse {