+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - `cycle` method for `FiniteIterable`s
 - `distinct` method for `FiniteIterable`s
 - `foldRight` method for `FiniteIterable`s
+- `reduceLeft` method for `NonEmptyFiniteIterable`s
+- `reduceRight` method for `NonEmptyFiniteIterable`s
 - `size` method for `FiniteIterable`s
 
 ### Changed